Replace rowMeans with apply(x,1,mean) (fix #14)
rowMeans provides slightly different results when run under valgrind.
According to Prof Ripley, this is possibly because "there will be more
stores to memory and so the extended precision of x86_64 CPUs will be
used less. I can confirm that via the use of flag -ffloat-store in
gcc/gfortran...".

This causes potential for TTR unit tests to fail for functions that call
rowMeans. Using apply with mean allows the tests to pass, likely because
of mean.default's second pass for error correction.