Add checks for single-column input (fix #15)
Many functions implicitly expect a vector or matrix/data.frame with
only one column, and some of these functions called complied code that
would crash when passed a multi-column object.

Add a check to make this expectation explicit and throw an error if
the object has more than one column. Add tests to check for these
errors, since it would be nice for some of these functions to support
multi-column objects as inputs.