Clarify error message for 'invalid n'
The "Invalid n" error is not user-friendly. Provide a more informative
error that provides the current value of 'n' and the possible valid
values (based on the number of observations in the actual data.

Add details about valid values of 'n' to the documentation. Fix
'obversations' typo as well.

Fixes #32.