Clarify invalid exact.multiplier error message
The "Invalid exact.multiplier" error is not user-friendly. Provide a
more informative error that provides the current value of
'exact.multiplier' and the possible valid values. Add details about
valid values of 'exact.multiplier' to the documentation.

Also flip '0 > exact.multiplier' to 'exact.multiplier < 0' so it is
more like the check for 'n', and less confusing to future me.

See #32.