Fix notation in documentation of volatility function
The volatility function in volatility.R includes a few equations under
"Details". Ideally, naming convention of the R functions and the
equations should match for better readability. The number of periods
per year was named N in the function but Z in the equations. The
comments in the volatility.R file were changed to make documentation
use the former. Commit includes the .Rd file made by roxygen2 5.0.1.