Clarify MACD and EMA documentation
The comment in the *Note* section of ?MACD said "the function originally
used exponential constants", which users may confuse to mean that
MACD() itself is or was using some constant exponential value. Change
the comment to reference the original _formula_.

Add a note to ?EMA to explain how the result is initialized using the
sample mean at period 'n'.

Thanks to Naresh Vadrevu for the suggestions.