Move rolling percent rank from Fortran to C
The .Fortran() interface has forced duplication of all inputs for
quite awhile now, and recently started to require registration of
native routines.  Moving from .Fortran() to .Call() avoids both of
these issues.

Note that this changes the behavior of function when cumulative = TRUE.
Previously, the first observation was always 0. Now it is set equal to
the exact.multiplier argument, which is more consistent with the
function's behavior in other cases.