Use price+index struct and improve naming
Track reference and inflection prices+index in a single struct to
improve code organization, and make it easier to copy the inflection
data to the reference object if the trend reverses.

Use more descriptive names for local and extreme maximums/minimums.
Also use descriptive names for input parameters and result object, and
name SEXPs with a leading underscore.