Correct/add colnames to CMO, RSI, and SAR output
The output of CMO was an unnamed matrix, when it should have a column
name of 'cmo'. The output of RSI was a matrix with a column name of
'EMA' (or whatever MA was used) instead of 'rsi'. Both of these are
fixed by checking the result for dims and 1 column, then setting the
appropriate column name.

SAR returned a vector, not a 1-column matrix. Allocate a matrix in the
C code and set its column names before returning from the R function.

Fixes #53, #54, #55.