Use try.xts/reclass and call ema() C function
Call ema() C function to avoid 2 calls to the EMA() R function. This
requires adding the try.xts() and reclass() calls to DEMA(), since
they were being done by EMA().

Also add other checks for the value of 'n' and the number of columns
of the input that were done in EMA().

Set 'n' to NULL if it's missing and 'ratio' is not. The C code will
calculate the value of 'n' for the given 'ratio'.

Fixes #69.