Update LaTex Yang-Zhang equation for 'k'
The ability to specify 'alpha' for the Yang-Zhang calculation was
added in commit eb0fd34332fca1cdef34849404dd4d249785bc76. The docs
were updated in commit 697012f51eb1e5941f693e9334dfefbb2ae00c88 to
provide LaTeX and text versions of the equations, but the Yang-Zhang
equation for 'k' was not consistent.

Make the documented LaTeX version of Yang-Zhang equation for 'k' match
the code.

Fixes #72.