Add cumulative tests, fix off-by-one errors
The tests failed until the off-by-one errors were fixed.

Add 1 to 'i' when passed to ttr_median() and ttr_mean() for the value
of 'n', since 'n' is one-based and 'i' is zero-based. We also need to
loop from 'j = 0' to 'j = i'.

Remove the 'window' object for the cumulative branch of runcov(),
since only ttr_mean() is called and it does not mutate its first
argument like ttr_median(). Also update the denominator to change
with 'i' since the window is expanding.