Add warning to DPO about using future data
The detrended price oscillator uses a centered moving average, which
looks at future data. Therefore this indicator should not be used in
systematic trading rules. Add a warning to that effect.

Remove the links to MetaStock and FM Labs because they are not
particularly clear, and my not be internally consistent.

Fixes #47. See #1.