add optional, more accurate mode to function runSum
If 'x' contains a mixture of numbers very different in size, the existing, fast
version becomes inaccurate; this can lead to small negative running sums on
all positive input data, which can in turn lead to various secondary errors
in other TTR functions (VWAP, CMF, DVI, MFI, rollSFM, and volatility) built
upon runSum.

Example: If some of the OHLC time frames have almost no price movement, the
corresponding values passed to runSum are very close to zero in function
volatility with Rogers Satchell algorithm. If such small numbers are often
interleaved with bigger numbers, the existing optimized implementation of
runSum will produce slightly bigger or smaller running sums on the long run.
This can end up with small negative values such as -0.1e-15 being passed to
sqrt, which is undefined.

Fix by adding an optional parameter 'accurate.instead.of.fast' with default
FALSE. When TRUE, not the existing, fast, incremental calculation of the
running sum is used, but the running sum is calculated separately for each
window position, which is slower by a factor of 'n', but accurate.