Ensure 'ratio > 0' in EMA(), ZLEMA()
Segfaults were possible when 'ratio = 0.0'.

EMA() would segfault when 'ratio = 0.0' and 'n' was not specified,
because 'n' is calculated using 'ratio' and would be NaN in that case.

The same was true for ZLEMA(), but it would also segfault if both 'n'
and 'ratio = 0.0' were specified. Because the C code calculates the
number of look-back periods as 'lag = 1 / ratio'. And we can't subset
the input vector by NaN.

See #95.