Handle NA in WMA() C code
WMA() did not return an xts object when 'x' is xts with leading NA,
and 'wts' was the default `1:n`. This was because calling na.omit()
caused 'x' and 'ma' to have different lengths, which caused reclass()
to return 'x' as-is.

Remove *both* calls to na.omit(), and the code that added NAs to the
'ma' object after the if/else statement to handle the different 'wts'
cases.

The C code only needs to find the first non-NA value. It doesn't need
to calculate the first result, like it does for the recursive MA
functions.

Thanks to Cory Fletcher for the report!

Fixes #96.