Fix cumulative median when input has leading NA
The calculation was wrong because 'window' is a copy of 'x' and we
always sorted from the beginning of 'window' to the current
observation. This caused  incorrect results when 'window' contained
leading NA.

Now we only consider observations between the first non-NA element and
the current observation.

Add an argument to ttr_median() to specify the first observation to
consider when sorting. This is consistent with the R_qsort() function
arguments. Update the calculations to determine if the number of
observations to be sorted is even, and the midpoint of the window.

Fixes #93.