Use NASDAQ FTP to get stock symbols
Users have been getting timeout issues and errors using the URL:
https://old.nasdaq.com/screening/companies-by-name.aspx

Switch to the FTP URL: ftp://ftp.nasdaqtrader.com/SymbolDirectory/.

Keep the column names from the old data set, but set all the values
to NA. Also keep the columns in the same order, and append the new
columns to the end of the output.

See #98.