Add leading NA if n > 1 and cumulative = TRUE
runPercentRank() would not pad the initialization values with NA,
which was not consistent with the other running functions in the
package.

It was simple enough to remove the if(cumulative) n = 1 from the
C code. But we also needed to update the if statement in the R code
that set all the results to the exact multiplier if n = 1. Because the
window expands to be > 1 in the cumulative case.

Fixes #73.