Fix stoch() when fastK is Inf
It may be possible for fastK to be Inf in very rare cases. I could only
reproduce this if the Close is > High and High and Low are equal. But
that is a data error. I fixed anyway because there may be cases I can't
think of at the moment.

Check for any non-finite values in fastK and replace them with 0.5.
Make sure not to replace any NA, because leading NA need to propagate.

Thanks to GitHub user cjuncosa for the report.

Fixes #52.