Fix MFI() money flow is always zero or positive
The denominator of the money ratio will be zero if there is no negative
money flow for 'n' consecutive observations (e.g. during a strong
up-trend), which causes the money flow index to be Inf. Set the money
flow index to 100 in this case.

And the money ratio will be NaN if there's no money flow for 'n'
consecutive observations (e.g. if there are no trades), which causes
the money flow index to be NaN. Set the money flow index to 50 in this
case.

Thanks to GitHub user jgehw for the report, reproducible example, and
suggested patch.

Fixes #81.