Use seq_len() to handle cases with no NA
It's a bit confusing to see `0:N` in R code, since indexing starts at
1. That was done here to handle the case when `n-1+NAs` was zero. That
worked because the 0th element doesn't exist, so nothing was set to
NA.

It's generally safer to use seq_len(), seq_along(), etc. instead of
`1:N` because the result is `c(1, 0)` when N=0. That's not intuitive
and often leads to bugs, which happened here (#108).