Fix runcov() starting point and initialize means
runcov() was starting one observation after it should have. And it did
not initialize the two series' means when they started somewhere other
than the first observation.

Move the 'first_i' calculate before the if/else for 'cumulative' so we
can use it in both. And always set the first element to NA because the
variance of one observation is not defined.

See #121.