Fixed Minor Bugs
The following MA-functions didn't add a column name as the SMA-function does:

HMA
ALMA
WMA
EVWMA

The WMA were missing a reclass before the if-statement, and were therefore giving dim = NULL, which also affected the HMA.

By fixing WMA, the ALMA and HMA got fixed as a by product by also adding an if-statement similar to the rest of the MA functions.

The EVWMA function were wissing a reclass before the if-statement, so it never got evaluated inside the if-statement.