Remove look-ahead in SAR
Calculating the sd() of the high-low using the entire series causes
look-ahead bias. Instead, use the sd() of the first high and low. Yes,
this is a bit silly, but it provides a better starting value than
setting the first SAR value just above the first observation's low.

This is a breaking change.

Thanks to @cybe-rstar for the report!

Fixes #135.